Display banner when running hh and there are unsaved changes in the editor
Summary:
Show a warning when the user runs `hh` on the command-line, but there are unsaved changes in the editor. This is to fix the linked task and make it clear to people why mysterious phantom errors may be appearing in the results of `hh`.

There is another diff out to add a timeout to Nuclide, but honestly, I don't know if it's worth trying to complete that and dealing with all the edge cases and then losing the ability to reconnect to your devserver after a weekend. I think this should be sufficient to address the real problem we're having, which is users confused about the source of the phantom errors. They can surely figure out that something is wrong from there.

Reviewed By: ljw1004

Differential Revision: D4708413

fbshipit-source-id: 4d5287c949d698f86a6e4e97622ce8642745d12f