Better global dce
Summary:
Dce finds stack locations that are "unused", and kills all the associated instructions.

eg

  String "foo"
  PopC

would be removed. Similarly

  String "foo"
  Null
  SetL 1
  PopC
  PopC

would still remove the String/PopC since nothing in between depended
on the string being on the stack.

but

    String "foo"
    CGetL 1
    JmpNZ L
    Null
    SetL
    PopC
  L:PopC

Would not remove the String/PopC because there's control flow in the middle.

This adds tracking for the dceStack across blocks, so that we can do
proper global dce.

Reviewed By: ricklavoie

Differential Revision: D4696844

fbshipit-source-id: 09e3b58f9ed70fc6dc86752320992bf66b1e4061