Add DceAction for cases where we can't just delete an instruction
Summary:
Sometimes we have a long chain of instructions where all can be
deleted as long as the first one can; or multiple paths to the same
pop, where none can be deleted unless they all can. This lets us
insert PopCs right after an instruction that can't itself be removed,
when it looks profitable to do so.

As proof of concept, use it for UnboxRNop.

Reviewed By: ricklavoie

Differential Revision: D4703563

fbshipit-source-id: 75cd2b59a28cd246193b3b28b5b30fd933b64a7f