Deleted instructions don't do anything
Summary:
Even after D4716282 I was seeing SetLs that appeared to be dead, but
weren't deleted. I noticed that CGetL called AddLocGen even when it
was going to be removed - so stop doing that. Auditing the rest of the
AddLocGens turned up one more case.

There are still a few places where we are probably going to delete the
instruction, but have to mark its effects anyway, just in case.

I have more dce diffs that will hopefully give us a way to deal with
that.

Reviewed By: ricklavoie

Differential Revision: D4728909

fbshipit-source-id: c33e812fe4d59090e5bcbcd750b433132154ff28