Implement legacy parsing for darray and varray
Summary:
Introduces syntax and parsing for `varray` and `darray`.

`varray` is a vector-like array. It is the same as `array<T>` for some `T`. KendallHopkins needs this in order to migrate `array`s used like vectors to `vec`s.

`darray` is a dictionary-like array. It is the same as `array<Tk, Tv>` for some `Tk` and `Tv`. KendallHopkins needs this in order to migrate `array`s used like dictionaries to `dict`s.

Reviewed By: dlreeves

Differential Revision: D4684529

fbshipit-source-id: c5072487a26ab0253a71579a4a55de6cc1821fb4