Share exit blocks for all initial guards in the tracelet selector
Summary:
We can create just a single exit block upfront and set it as the "guard-failure"
block, thus making all the guards branch to it instead of creating a different
exit block for each guard.  This avoids creating a bunch of exits, particularly
for functions with a large number of locals, thus speeding up the tracelet
region selector.

Reviewed By: paulbiss

Differential Revision: D4733191

fbshipit-source-id: 80ac6988a490d3743abe7cfefa53ac4b8e029ca3