Allow hhbbc to run on previously optimized repos
Summary:
It wasn't allowed, because it didn't work. The main reason was that
AssertRATL didn't mark its local as read, so we could eliminate it,
but still try to spit out the AssertRATL.

This itself is an indication of a problem with hhbbc - because the
local should have been eliminated first time around - and is in fact
the issue I was investigating when I started trying to run hhbbc
twice.

To fix the AssertRATL issue, I added the read; but then the final
bytecode ends up with extra AssertRATs in it. It would be nice to just
optimize redundant AssertRATs away, but there's currently no
RepoAuthType->Type converter (and its not really going to buy us
anything for now), so I'm also stripping them out at parse time.

Having done this, I modified test/run to run hhbbc twice, and ran all
the tests; which found a number of minor bugs - none really related to
running hhbbc twice, except that again, hhbbc is failing to follow
through with certain optimizations in certain situations. I'll follow
up on those next.

Reviewed By: jano

Differential Revision: D4734124

fbshipit-source-id: c8288dd1e3c6d6f1449823eda45ce1f72f83ea13