Fix some inconsistencies in array_like_elem
Summary:
Given an unknown key, an ArrLikeMap would return the union of all the
values and InitNull; if all the values were similar (eg all Ints), we
could get a reasonable type. If the ArrLikeMap got promoated to an
ArrLikeVal (because we resolve the exact values of all the elements),
we just returned InitCell. This could result in (for example) a return
value being determined as one type early on, and TInitCell later.

This makes them both behave the same way. At the same time, it fixes
them to only return the union of the elements whose keys *might*
match; and makes them work with Hack arrays.

Reviewed By: paulbiss

Differential Revision: D4736144

fbshipit-source-id: a71d241dd5491f921e4b01573fd08b5bb8caa25f