Fix another issue with const-prop (take two)
Summary:
When we constprop, the resulting type will always be a static type,
but typically we'll have a non-static type. eg concat of two literal
strings will produce TStr, since that's what we'll get from a concat
at runtime - but once we propagate constants, we'll have a literal
result too.

During the analysis phase, we don't actually modify anything, so we
could end up inferring a return type as a non-static type, and then
never fix it up. Running hhbbc a second time would see the literal,
and infer the correct return type. This fixes it so we get the correct
return type first time around.

Reviewed By: paulbiss

Differential Revision: D4736147

fbshipit-source-id: d07e6347a6884e9c053bfd67fab9f57e9ff05e00