Start emitting collection literals in attributes
Summary:
We need to emit collection literals in attributes and property initializers.  This will be necessary for its own sake, and also needed to get memoization codegen right.

In this diff, we now support zero-length dictionaries in attributes, and that's all.

In upcoming diffs: dictionaries of more than zero length, collections other than dictionaries, properties with initializers.

Reviewed By: oulgen

Differential Revision: D4724051

fbshipit-source-id: d150138364cb46d4be194a7cc3bb35d79e882249