Handle recursively defined class constants
Summary:
We should be raising an error if we detect a class constant is recursively
defined. Instead we just loop endlessly and crash when we run out of stack
space.

Fix this by keeping track of what class constants we're currently defining. If
we attempt to define a class const we're already defining, then raise the
error. Use a special marker array as the value in the class constant array to
indicate the class constant is currently being defined.

HPHPC has a similar issue when it attempts to constant propagate the class
constant values. For recursively defined class constants, it would loop
endlessly because the value can never be flattened. The fix for this is a little
more complex. For each class constant, gather any class constants which might be
part of that class constant's definition. Then, expand out all the transitive
class constant dependencies. At this point, if a class constant has a dependency
on itself, it is recursively defined and and should be marked dynamic. We won't
attempt to const propagate the values of such dynamic class constants.

Reviewed By: markw65

Differential Revision: D4740254

fbshipit-source-id: a8d7e7967f4c677d40e43fea9509baf2f5463751