Allow PHP and Hack arrays inside constants
Summary:
Add support for both PHP and Hack arrays inside constants, using both the
define() and const syntax. This applies to both top-level and class constants.

Most of the support was already there. We just needed to remove some checks in
the emitter, and have Variant::isAllowedAsConstant() to allow arrays and check
their contents recursively.

Closes: https://github.com/facebook/hhvm/issues/4277

Reviewed By: markw65

Differential Revision: D4730483

fbshipit-source-id: 6ca45d45c369634925f5f93fbf7856accb718ad6