Fix LocalRange immSize decoding & interpOne(Silence)
Summary:
immSize() could incorrectly decode the second IVA field of
a LocalRange immediate, if the first one was 4 bytes.

irgen-interpone for Silence was testing the wrong subop (OA)
immediate number, (0 should be 1), and then passing the local
id to setImmLocType() intead of the immediate index of the local id.
(instr.imm[0].u_LA should be just 0).

Reviewed By: ricklavoie

Differential Revision: D4735174

fbshipit-source-id: ec61472604276b50f939a4001ad0b85208c714ec