Add an hhbbc2 option to test/run
Summary:
This adds the option --hhbbc2 to hphp/test/run. It implies --repo, and
it causes us to run hhbbc, dump the hhas, run hhbbc again, dump the
hhas again and compare; if they're different, the test fails there,
otherwise it runs the test as normal (against the doubly optimized repo).

Reviewed By: ricklavoie

Differential Revision: D4750683

fbshipit-source-id: c444a80785c1c48b36efec7a12ae00d487e9ef0e