Fix some issues with strength-reduction and constant propagation
Summary:
When we strength-reduced, we would only constant propagate the last
replaced bytecode; fix that by constant propagating as we go.

If one of the elements of an impl sequence got strength reduced, the
entire thing got strength reduced. This is a problem because eg
FPushClsMethodF uses impl(FPushClsMethod), since as far as hhbbc's
analysis is concerned, the two are equivalent. But if we decided to
strength reduce FPushClsMethod to FPushClsMethodD (we don't currently
do this optimization), we would then actually replace the
FPushClsMethodF with FPushClsMethodD, which would be wrong.

This fixes impl so it never strength reduces as a side effect.

Reviewed By: ricklavoie

Differential Revision: D4735190

fbshipit-source-id: 2be119207b22a4c4a23d933a24efa6411e000c4a