Fix an optimality issue with global dce
Summary:
Global dce was designed to compute global liveness by running
analysis once, and then iterating to convergence. To do so, it had to
assume everything was live out during the analysis pass. This was
slightly conservative in the cases where liveness affected the usage
of stack slots, or the liveness computation itself. I recently updated
the algorithm to also compute liveness of stack slots, and switched
to repeating the analysis but continued to compute liveOut using the
same approach. Its much better to track liveOut directly, and let the
local analysis propagate it back to liveIn.

This produces much better results for the stack slot usage
computation, where the usage is often dependent on the liveness of
a local.

Reviewed By: ricklavoie

Differential Revision: D4735445

fbshipit-source-id: 4f20c0a20b4140ccf68bf9ff3b277c393fdc700b