Add more function specific tracing
Summary:
hhbbc's --trace option can be used to specify a set of methods that
you want to trace, which can be useful when debugging issues in large
builds.

For my debugging, I found that the trace level (2) was not providing
enough information, and that I also needed to see what was going on in
cfg-opts, and dce.

Reviewed By: ricklavoie

Differential Revision: D4754778

fbshipit-source-id: 47fb5947a414c56d895b80aa1d9c17a241535e85