Fix block merging in the presence of factoredExits
Summary:
Since we were counting factoredExits in the number of successors, it
was impossible to merge a block with factoredExits into its successor.

Once we fix that, we're already checking that they have the same exnNode,
and the same section, so they must have the same factoredExits, so we can
ignore factoredExits when computing numSucc, and remove the code that
merged the factoredExits.

Reviewed By: ricklavoie

Differential Revision: D4751353

fbshipit-source-id: c8adb80d66afde88bc739888617ffa68390b5829