Introduce fmt_node IR in hackfmt
Summary:
Add a new IR between the Hack AST and chunk groups. This change makes the `transform` functions in hack_format.ml purely functional, returning a declarative description of how a node should be formatted.

Removes `transform`'s dependency on the chunk_builder, and the chunk_builder's dependency on the Hack AST.

Reviewed By: eshrews

Differential Revision: D4723095

fbshipit-source-id: 37bfdad3898b6e200843e30b85a6138056491b92