Don't split before comments
Summary:
Previously, we were adding a simple split before comments (if unsplit). This forced the current argument rule to split, which interacted badly with skip_parent nesting.

With this change, we no longer add a split before comments, so trailing comments will stay on the same line as the code they are annotating.

Depends on D4723095

Reviewed By: eshrews

Differential Revision: D4741258

fbshipit-source-id: 3a025af730e2b3d445866fa7fd257d93a9134047