Allow resources as constants
Summary:
PHP allows resources as constant values (not just the predefined ones like
STDIN), though they must be initially created using define(). Like so:

define('FOO', xml_parser_create());
const BAR = FOO;
class A {
  const BAZ = FOO;
}

We don't, so fix that. This means that we cannot automatically make anything
defined as a constant static since resources cannot be static. However, we
shouldn't be doing that anyways as it can cause memory leaks if the constant has
a different value for every request. Instead ref-count the constant value as
normal. This exposed a couple of ref-counting bugs with constants which didn't
matter previously because everything was static. Namely, LdCns was marked as
producing a ref-count, but never incremented it.

Reviewed By: markw65

Differential Revision: D4755663

fbshipit-source-id: 022b54df472c9fba8e0da7292e9cadfa39ae13ef