Implement typing for darray and varray
Summary:
`darray` and `varray` now have all of their typing constructs in place. Specifically for purposes of subtyping and unification, we handle `darray` the same ways that we handle `AKmap`, and we handle `varray` the same ways that we handle `AKvec`. Sicne we keep `darray` and `varray` distinct at the `typing_defs` level, however, we are still able to offer the correct suggestions for users that would use `darray` and `varray`. This capability would be lost if we simply lowered the type to an `AKmap` or an `AKvec`!

This design is also ideal moving forward. Since the subtyping behavior of `darray` and `varray` is isolated into `typing_subtype.ml` and `typing_unify.ml`, in the future when we want to increase the strictness of `darray` and `varray` subtyping, we will be able to do this //without// touching or risking changes to any of the existing array/vector/map code.

Reviewed By: eshrews

Differential Revision: D4691149

fbshipit-source-id: c7cd6e753fb6593ba683e898d4acc105ad3df89a