Fix AsyncMysqlConnectionPool sweep
Summary:
It seems that in the conversion to HNI this flag to disable SWEEP was added, and that seemed to have prevented this object from destruction.
I added the sweep call and changed the flag. The sweep call also shouldn't be necessary, but I want to explicitly call shutdown.

Reviewed By: paulbiss

Differential Revision:
D4773683

Tags: scgraph

fbshipit-source-id: 7e60f7a8416f1aeb574198e93005aa1fbdf76860