Fix HPHPC handling of foldable functions that can return Hack arrays
Summary:
When HPHPC folds a foldable function to a constant, and that constant
is array-like, it assumes that its a PHP array. This can lead to Hack arrays
mysteriously becoming PHP arrays in certain contexts. Instead check for the
actual type of the array and convert it to the right AST node as appropriate.

Reviewed By: markw65

Differential Revision: D4771956

fbshipit-source-id: e546617223f6c12eb211575405450febdb604a6d