Make "stack" member of CallEffects more precise (take two)
Summary:
The "stack" member of CallEffects indicates which stack locations a
call may read or write from. However, all the call instructions set this to
AStkAny, which is correct, but not optimal. Change this to properly indicate the
stack locations which might actually be read or written from. A call will
generally read/write to its act-rec, its parameters, or any stack slot below the
act-rec or params.

The original attempt at this had to be reverted because of a bug in the logic I
added to avoid keeping tmps alive across calls. A TrackedStore was being marked
bad in a place where the resolution logic wasn't expecting it. This updated
attempt has that bug fixed.

Reviewed By: markw65

Differential Revision: D4778253

fbshipit-source-id: 9a56531c6facd9b0e6630a2ee7751e75e835eb79