Implement full-fidelity parsing for darray and varray
Summary:
Introduces syntax and parsing for `varray` and `darray`.

`varray` is a vector-like array. It is the same as `array<T>` for some `T`. KendallHopkins needs this in order to migrate `array`s used like vectors to `vec`s.

`darray` is a dictionary-like array. It is the same as `array<Tk, Tv>` for some `Tk` and `Tv`. KendallHopkins needs this in order to migrate `array`s used like dictionaries to `dict`s.

Differential Revision: D4724716

fbshipit-source-id: e4f4409cc387090ec3b0fcb41d5d1fadc8c483d8