Implement full-fidelity parsing for optional field types
Summary:
Introduces the syntax for optional field types into the full fidelity parser.

A field now has a `field_question` identifier. This will either be a `Question` in the case of an optional field, or a `Missing` in the case of a mandatory field.

I've added a `TODO` for supporting `hackfmt`. I will accomplish this in a subsequent diff.

I've updated the pretty printer in `full_fidelity_pretty_printer.ml`.

= Questions

ericlippert mentioned that there were two pretty printers; however, I could not find what the other one was the needed to be updated.

ericlippert also mentioned:

> We need to make a spec issue to update the spec appropriately. (The spec is on GitHub.)

Is a "spec issue" simply a Facebook task, or is it a more specialized process? Let me know if there is immediate action that I should take on this diff.

Is there a "lowering" step that I have to implement in order to translate the full-fidelity AST into one that looks like output from the legacy parser?

Reviewed By: ericlippert

Differential Revision: D4595482

fbshipit-source-id: e6958b39f61dcfb5e142126ede673e19be44aacb