Allowing `noreturn` in type specifiers
Summary: The full fidelity parser did not allow `noreturn` in arbitrary type hints; only in return types of functions and in generics (because `Awaitable<noreturn>` also had to be permitted). There are tests in our test set that expect a `Naming[2063]` error for the use of `noreturn` in the inappropriate location. Since naming does this check, the parser should be as permissive as possible, to help users of the parser (LSP, HackAST, etc.) to deal with such scenarios.

Reviewed By: ericlippert

Differential Revision: D4721699

fbshipit-source-id: 50fe2fe8a54aa4cec8a84579ab780d606702cfc3