Fix tc-print with perf counters
Summary:
I recently killed the support for instrumentation-based profile counters, but
missed one place.  This was triggering an assertion when trying to use tc-print
with profile counters.  This diff fixes the problem, and it also replaces the
assert with a static_assert, so that such errors can be caught at build time.

Reviewed By: paulbiss

Differential Revision: D4781500

fbshipit-source-id: a985c8dfdc630658b2056c9f8f4f050ea1ecd240