Exact-scan Waithandles with type_scan
Summary:
Add type_scan support for WaitHandles. the type index is stashed in a
new field c_WaitHandle::m_tyindex, and used at scan time to invoke the
appropriate scanner for the concrete waithandle type.

in AsioBlockable, refine m_bits into a union with AsioBlockable* to more
accurately reflect the existence of a pointer. And scan the pointer.

in AsioBlockableChain::unblock(), preserve the parentChain linked list as
we pop and unblock one WH at a time. in unblock() and removeFromChain(),
null out the removed element's nextParent field to avoid dangling pointers
which can cause false positives with conservative scanning.

Reviewed By: jano

Differential Revision: D3053017

fbshipit-source-id: aa3ee2d25bcc32629fc98956ce56eeae14232172