Handle trailing comma before final item trailing trivia
Summary:
If the final item in a comma list has a trailing line comment preceding a trailing comma, hackfmt would insert that trailing comma into the line comment, breaking idempotence.

Adding the trailing comma before the trivia ensures that this doesn't happen. It also means we produce consistent output when the comma is present or missing.

Depends on D4746056

Reviewed By: eshrews

Differential Revision: D4790466

fbshipit-source-id: 86092add0e09399c32c44dbf89bb5245435fb816