Fix bugs in foreach codegen
Summary:
I noticed that the key and value were passed to IterInit/IterNext in the wrong order,
on further inspection the labeling and branching was also wrong and IterBreak/IterFree
were not called for break or return inside the loop. This gets a few new tests passing,
some of the new passes are from features that were implemented.

Reviewed By: andrewjkennedy

Differential Revision: D4802875

fbshipit-source-id: cd44d03ade9ab87b4bfc9dfb6307104a14ac2262