Added `Mphp` file mode to offer better front-end support for PHP
Summary: With the introduction of HackAST and other tools built on the Full Fidelity Parser, we're getting bug reports for `<?php` files that do not parse. To have a better granularity to still support PHP in the front-end (parser), but not the back-end (type-checker) - and later maybe also in the back-back-end (codegen) - it helps to know that we're talking about a PHP file.

Reviewed By: oulgen

Differential Revision: D4738731

fbshipit-source-id: 36780e850bef5be37aa7b25136bba7d08b6bfc0f