Add support for HackArrCompatNotices in HPHPC and HHBBC
Summary:
Up until now we've been trying to keep HackArrCompatNotices effects isolated to
runtime, so we could preserve bytecode compatibility and turn it on and off at
will. However, this doesn't really work as HPHPC and HHBBC can perform
optimizations which prevent us from warning at runtime in every circumstance. We
can remove these optimizations, but then we must pay the cost everywhere even
with the option disabled. Add support for the HackArrCompatNotices option to
HPHPC and HHBBC and only perform the pessimizations when its enabled. Note that
its legal to compile the repo with HackArrCompatNotices enabled, but disable it
at runtime. However, the opposite is not allowed (HackArrCompatNotices at
runtime, but not at repo build time).

Reviewed By: markw65

Differential Revision: D4805376

fbshipit-source-id: 7b23a33bd09f958dcea097756d3e01449fb1cbf4