Kill tech-debt for h2tp
Summary:
We've had lengthy discussions about the maintenance of `h2tp` and `hh_matcher`. Considering that

  1. They have not been maintained and have significant features go unsupported for a while now;
  2. There are better alternatives for `hh_matcher` now that we have a full fidelity parser;
  3. Given that we want to start using the full fidelity parsers as the new front-end for quite a few of our tools, these deprecated and bit-rotting products pose too high a cost to keep;

these tools will now be removed. This diff removes `h2tp` and its associated tests from the code base.

The purpose of this diff is to stack the deletion diff for `hh_matcher` on top of this and ship them together. This is why I've simply stubbed out the `h2tp` facilities in `hh_matcher` and simply took out the three failining `hh_matcher` tests.

Reviewed By: andrewjkennedy

Differential Revision: D4802591

fbshipit-source-id: fc62450b593ec42337c3e0549846bb1c6db7177e