Propagating correct ParserOptions in the lowerer
Summary:
Most of the old parser plumbing just used `Namespace_env.empty_with_default_popt`, but recent discussion lead me to believe forgetting about the parser options wasn't going to be as inocuous as it seemed. Here now, the plumbing to actually pass the parser options around (the parser itself didn't actually do anything with them).

Taggin oulgen as reviewer, because he seemed to be most up on the how and why of parser options.

Reviewed By: oulgen

Differential Revision: D4635534

fbshipit-source-id: 37da3ca1d27606efbdee5807f0125b56d7b68686