implement textDocument/documentSymbols
Summary:
This is the implementation of documentSymbols, which we'll use for outline.

Note: here we flatten the symbol TREE into a flat list where each node mentions the string name of its container. From this, we can reconstruct a tree so long as no container itself contains a container of the same name. As far as I can tell, that never happens in Hack, so we're okay. When we use LSP for other languages we'll have to figure out a solution.

Reviewed By: dabek

Differential Revision: D4776178

fbshipit-source-id: e8188fd2c2849d80da5882842dbd422bd3e35966