Don't acquire the codeLock when handling bind {jmp,jcc,call,addr}
Summary:
Introduce per SrcRec and per ProfTransRec locks and stop using the codeLock to
protect the bind{Jmp,Jcc,Call,Addr} helpers. The lock used in SrcRec replaces the
atomic guard (and is the same size).

Removing the overuse of the code and metadata locks from tc-recycle.cpp is a
logical follow up.

Reviewed By: ottoni

Differential Revision: D4704952

fbshipit-source-id: 69750b5e0b1ac0ccc3d3d42c5f49dd66bca0d2ec