Use a worker thread for reclaiming Funcs/SrcRecs/Translations
Summary:
This way we won't fill the treadmill with retranslate jobs and tie up all of the
request handling threads. Doing this on the treadmill was always suspect
anyway, particularly when the entire process was gated on the code lock.
This diff also addresses the lock contention issue by avoiding the code and
metadata locks where possible. In a follow up I may add a separate lock for
the DataBlock allocator to avoid holding the code lock entirely.

Reviewed By: ottoni

Differential Revision: D4724004

fbshipit-source-id: 4f7f5f018d6b51d5d086652373bb45ce57d6c3be