implement textDocument/references
Summary:
This implements the FindAllReferences part of LSP.

Here is the corresponding Nuclide diff: D4783358

Screenshot with the status quo:
{F66277318}

Screenshot after this diff, using LSP:
{F66277335}

Note that the status quo uses the fully-qualified symbol name in both the editor tab name (where it is far too long!) and in the title of the find-results window.

LSP doesn't offer up the symbol name at all, neither brief nor fully-qualified. I wanted to write Nuclide to work with any LSP client, so I wanted to see how far I could get by just with LSP as is. (Indeed, VSCode highlights the symbol with existing LSP clients). So on the Nuclide side, it derives the symbol name by using the contents of the text buffer at the specified range.

Reviewed By: dabek

Differential Revision: D4783324

fbshipit-source-id: f91fa90896f8fb0f5298cfb24190defb8fa8f66f