Add `locals` set to CallEffects, use it in store-elim
Summary:
Remove the hard-coded assumption in store-elim that all locals are read in a
call instruction. Instead, add a new alias-class in CallEffects which gives the
locals which may be potentially read by that call and use that instead.

As part of adding the `locals` set to CallEffects, tighten up the alias-class of
what locals can actually be read in a call instruction. If
EnableArgsInBacktraces is enabled, only union the frame locations which
represent parameters, rather than the whole frame. The logic also failed to
consider the 86metadata variable, so consider that as well.

Certain functions can read arbitrary locals from the caller's frame, so add a
new "readsLocals" attribute to the Call IR instructions, which analogously to
"writesLocals" indicates this. If "readsLocals" is true, the call is assumed to
be able to read any local.

Load-elim was assuming that CastStk and CoerceStk's load alias-class always
contained an exact stack location, which is no longer always true. Instead
compute the stack location directly from the instruction.

Reviewed By: markw65

Differential Revision: D3963830

fbshipit-source-id: a593eab649e78cd845e5fb1b2119b86cce179ced