implement textDocument/documentHighlight
Summary:
This is the LSP request for highlight symbols within document.

The code for ide_range_to_lsp is ugly. I think that's a fair reflection of the OCaml ugliness of warning 45 "this shadows the label line" and didn't try to sugar-coat it.

Ultimately I think we expect the ide_ stuff to go away, and Hack internals will stick solely with Hack datastructures, and LSP stuff map those to the LSP datastructures.

Reviewed By: dabek

Differential Revision: D4787436

fbshipit-source-id: 8e6758338b5770722b31b8058e5614e4e8e3609c