IDE_FIND_REFS can also report the definition location
Summary:
Does the definition of "foo" itself count as a reference to the symbol "foo"?

On the one hand, no of course not. On the other hand, when users do FindAllReferences, then despite the name of this action they generally *do* want to see the definition included amongst the list of results. That's the default in VSCode, and it's an option in LSP, and so we should offer it as well.

I'm plumbing through the option. I've only enabled it for LSP clients at the moment (not clientIde.ml since that's going away; not at the command line since I've not heard requests for that).

NOTE: this is my first time adding stuff into the Hack server. It felt like heavy-going. Have I approached this the wrong way?

Reviewed By: arxanas

Differential Revision: D4817275

fbshipit-source-id: c2dc28d063a0fc8a437e7d763dc1cd2d425c615c