implement textDocument/typeCoverage
Summary:
This provides type coverage.

All the Nuclide client wants to know is (1) what is the percent-covered so it can be reported in the status bar; (2) a list of uncovered ranges+messages to report throughout the file.

In the case of Flow, the flow server reports (1) percent-covered, (2) uncovered ranges without messages. Nuclide synthesises a message based on the name of the provider.

In the case of Java, the java server reports (1) percent-covered, (2) uncovered ranges with messages.

In the case of Hack, the hack server used to split the file contents up into individual strings and send each string along with its coverage level (covered, uncovered or partial). Waleed changed it to send ranges with their coverage levels. Nuclide had to write a Hack adapter which turns those "covered vs uncovered vs partial" ranges into a percentage number, and which filters it to display only the uncovered/partial ranges, and synthesizes a message based on whether it's uncovered or partial.

I wanted to write LSP as the lowest common demoninator of all of these. So: LSP has the server send percentage, and ranges to report plus a message for each of them. (i.e. pretty much the same as Java).

This means we take the Hack adapter code out of Nuclide, and rewrite it inside hh_client. It still has to do the same job of course.

I'VE MADE A FEW CHANGES in what ranges it reports, and what percentage-coverage it calculates. It still has the sensible property that "a fully typechecked file reports as 100%, and a not-fully-typechecked reports as less than 100%, and every time you add typechecking to something then it increases the number". But consider this example:

  array_slice($f, $g)

  // Internally Hack makes this judgment:
  UNCOVERED: "array_slice($f, $g)"
  COVERED: "$f, $g"

  // Previously Hack reported its judgment like this:
  UNCOVERED: "array_slice("
  COVERED: "$f, $g"
  UNCOVERED: ")"

  // Now Hack reports these ranges:
  UNCOVERED: "array_slice($f, $g)"

What users will see is a bit less visual clutter, since the trailing parenthesis didn't really warrant its own error report. And they'll get a truer representation of what Hack internally things. But they'll see a touch less precision.

Reviewed By: arxanas

Differential Revision: D4798946

fbshipit-source-id: 3d035d0e06966376f5ca3956302ca397000b78b8