Implement legacy parsing for unknown shape fields
Summary:
Introduces parsing for unknown `shape` fields. An unknown shape field is a field whose types are indicated as being not completely known:

  type ShapeWithUnknownFields = shape(?'a' => int, ...);

In this situation, `ShapeWithUnknownFields` allows for unknown types. As a result, the type for `Hshape` has been changed as follows:

```lang=diff
  Hshape of
  -   shape_field list
  +   {
  +     si_allows_unknown_fields : bool;
  +     si_shape_field_list : shape_field list;
  +   }
```

In the code, the new record type is named as `shape_info`.

Note that this diff does not update the changes needed for naming, which will prevent unknown shape fields from being valid Hack, until we're ready to roll this feature to a wider audience. The main change that will be needed for this to work is in `naming.ml`.

Reviewed By: jamesjwu

Differential Revision: D4624133

fbshipit-source-id: 97a866e8da1dd0c66d5d5a77be57d601f9e9ba28