Implement typing for unknown shape fields
Summary:
When defining a shape, its `shape_fields_known` attribute is given a value based on the following decision matrix:

| | No ... | ...
|---|---|---
| Experiment off | `FieldsPartiallyKnown` | Error
| Experiment on | `FieldsFullyKnown` | `FieldsPartiallyKnown`

All of the enforcement associated with `FieldsPartiallyKnown` and `FieldsFullyKnown` is already implemented in the codebase, so there is no additional work required beyond that. Please see the "Test Plan" for information on how the tests have been updated in the face of the new functionality.

= Important note

There is some unusual behavior (documented in the tests) regarding the behavior of `idx`. I'll highlight the situation here:

== Case 1

=== Source

```hack
type ShapeWithKnownAndUnknownFields = shape(
  'a' => int,
  ...
);

function foo(ShapeWithKnownAndUnknownFields $argument): void {}

function bar(): void {
  foo(shape('a' => "tingley"));
}
```

=== Type check

```
line 11, characters 7-11:
Invalid argument (Typing[4110])

line 4, characters 10-12:
This is an int

line 11, characters 20-28:
It is incompatible with a string
```

=== Commentary

We have a shape that supports unknown fields, `idx`ed for a field that doesn't exist, and got a type error. According to andrewjkennedy, we actually should //not// error in this case. Ultimately, we want to allow this at the type level and produce a lint warning instead. However, after chatting with dlreeves, we've decided that since this matches the current behavior of `idx` for `FieldsPartiallyKnown` shapes, it's OK to hold off this work.

== Case 2

=== Source

```hack
type ShapeWithKnownFields = shape();

function f(ShapeWithKnownFields $shape): void {
  Shapes::idx($shape, 'x');
}
```

=== Type check

```
No errors
```

=== Commentary

We have a shape that does //not// allow unknown fields. You can `idx` into any field of it. Why?

Shapes that do not allow unknown fields are marked as `FieldsFullyKnown`. These kinds of shapes allow `idx`ing into any field. It's just the way the world is. I chatted with andrewjkennedy about this. andrewjkennedy has confirmed that it is OK to not produce an error here (read more about the discussion in my post). Ultimately, we should produce a lint warning, but not doing anything is OK (from andrewjkennedy's perspective) for the meantime.

Differential Revision: D4635865

fbshipit-source-id: 610d96ac95bca25ccee5db4f062ee1222ae3146d