Refactor range formatting
Summary:
Keep track of the character range in the original file represented by each chunk.

This allows us to know what range of the original file is represented by any chunk group, enabling formatting the chunk group at a given offset or returning an expanded formatting range in terms of offsets in the original file.

This also allows us to expand ranges to chunk boundaries, which is necessary for range formatting of heredoc strings and range formatting at offsets other than line boundaries (if we ever choose to implement that).

Also modifies the line splitter to solve only the chunk groups in the range being formatted.

Reviewed By: eshrews

Differential Revision: D4828774

fbshipit-source-id: 1e74880df10d00f87944ed1195202c683898c44b