Add splits after leading operators
Summary:
Currently, when a formatting range ends after a trailing binary operator, and we change it to leading, then the remainder of the text in the binary operator's chunk would also be emitted (even though those tokens were not in the requested range).

This change ensures that the formatting range end will fall on a chunk boundary, preventing this failure for binary operators.

Ultimately, we probably should just expand the formatting range instead of splitting, but until we can do that, this will mitigate the problem.

Depends on D4829869

Reviewed By: eshrews

Differential Revision: D4839735

fbshipit-source-id: 8f4d03d6f54fe614fa4de3c78cdd80d185145d26